ReportsOrderandcarad

+++ AutomotiveSols.BLL/Models/AutoPart.cs
@@ -12,16 +12,16 @@ namespace AutomotiveSols.BLL.Models
 
         [Required]
         [Range(1, 10000)]
-        public double ListPrice { get; set; }
+        public int ListPrice { get; set; }
         [Required]
         [Range(1, 10000)]
-        public double Price { get; set; }
+        public int Price { get; set; }
        // [Required]
         [Range(1, 10000)]
-        public double Price50 { get; set; }
+        public int Price50 { get; set; }
        // [Required]
         [Range(1, 10000)]
-        public double Price100 { get; set; }
+        public int Price100 { get; set; }
        // [Required]
         public int CategoryId { get; set; }
